move `Display` for `Algorithm` to correct file
