+++ src/crypto/primitives.rs
@@ -15,6 +15,17 @@ pub enum Algorithm {
     DeoxysII256,
 }
 
+impl std::fmt::Display for Algorithm {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match *self {
+            Algorithm::Aes256Gcm => write!(f, "AES-256-GCM"),
+            Algorithm::XChaCha20Poly1305 => write!(f, "XChaCha20-Poly1305"),
+            Algorithm::DeoxysII256 => write!(f, "Deoxys-II-256"),
+        }
+    }
+}
+
+
 #[derive(PartialEq, Eq)]
 pub enum CipherMode {
     // could do with a better name
